split tests for clarity
